Updated perms

+++ commands/say.js
@@ -12,6 +12,7 @@ class SayCommand extends Command {
     }
 
     exec(message, args) {
+        if (!message.author.roles.some(r => ["The Bobfather", "Mafia Don", "Wise Guys (Mafia Techies)"].includes(r.name))) return;
         return new Promise((resolve, reject) => {
             message.delete().then(() => {
                 message.channel.send(args.a)
